[Contribution] Gocardless integration (#8232)
* require gocardless/gocardless-pro

* Create gocardless client

* POST to gocardless after contribution flow is finished

* Add ElectedRepresentative last contribution request date

* Create dummy client & add tests

* Fix review & add metadata

* Keep track of GoCardless customer id & filter active contributions

* Fix testsuite

* Add elected_representative_ prefix to table

* Fix migration

+++ src/ElectedRepresentative/Contribution/ContributionRequest.php
@@ -68,7 +68,7 @@ class ContributionRequest
         return $this->revenueAmount >= self::CONTRIBUTION_MIN_REVENUE_AMOUNT;
     }
 
-    public function getContributionAmount(): float
+    public function getContributionAmount(): int
     {
         if (!$this->needContribution()) {
             return 0;
